Moved suite_xml into helpers.make_build
