+++ corehq/apps/app_manager/app_schemas/session_schema.py
@@ -8,7 +8,7 @@ from corehq.apps.app_manager.util import is_usercase_in_use
 def get_session_schema(form):
     """Get form session schema definition
     """
-    from corehq.apps.app_manager.suite_xml.sections.entries import EntriesHelper
+    from corehq.apps.app_manager.helpers.make_build.suite_xml.sections.entries import EntriesHelper
     app = form.get_app()
     structure = {}
     datums = EntriesHelper(app).get_datums_meta_for_form_generic(form)
