+++ corehq/apps/app_manager/detail_screen.py
@@ -1,8 +1,8 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 from corehq.apps.app_manager.helpers.make_build import id_strings
-from corehq.apps.app_manager.suite_xml import xml_models as sx
-from corehq.apps.app_manager.suite_xml import const
+from corehq.apps.app_manager.helpers.make_build.suite_xml import xml_models as sx
+from corehq.apps.app_manager.helpers.make_build.suite_xml import const
 from corehq.apps.app_manager.xpath import (
     CaseXPath,
     CommCareSession,
