+++ corehq/apps/app_manager/helpers/make_build/suite_xml/post_process/menu.py
@@ -1,8 +1,8 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 from corehq.apps.app_manager.helpers.make_build import id_strings
-from corehq.apps.app_manager.suite_xml.contributors import PostProcessor
-from corehq.apps.app_manager.suite_xml.xml_models import Menu, Text
+from corehq.apps.app_manager.helpers.make_build.suite_xml.contributors import PostProcessor
+from corehq.apps.app_manager.helpers.make_build.suite_xml.xml_models import Menu, Text
 
 
 class GridMenuHelper(PostProcessor):
