+++ corehq/apps/app_manager/helpers/make_build/suite_xml/sections/fixtures.py
@@ -1,13 +1,13 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 from lxml import etree
-from corehq.apps.app_manager.suite_xml.contributors import SectionContributor
-from corehq.apps.app_manager.suite_xml.xml_models import Fixture
+from corehq.apps.app_manager.helpers.make_build.suite_xml.contributors import SectionContributor
+from corehq.apps.app_manager.helpers.make_build.suite_xml.xml_models import Fixture
 
 
 class FixtureContributor(SectionContributor):
     section_name = 'fixtures'
-    
+
     def get_section_elements(self):
         if self.app.case_sharing:
             f = Fixture(id='user-groups')
