+++ corehq/apps/app_manager/helpers/make_build/suite_xml/sections/menus.py
@@ -3,8 +3,8 @@ from __future__ import unicode_literals
 from corehq.apps.app_manager.helpers.make_build import id_strings
 from corehq.apps.app_manager.exceptions import (ScheduleError, CaseXPathValidationError,
     UserCaseXPathValidationError)
-from corehq.apps.app_manager.suite_xml.contributors import SuiteContributorByModule
-from corehq.apps.app_manager.suite_xml.xml_models import Menu, Command, LocalizedMenu
+from corehq.apps.app_manager.helpers.make_build.suite_xml.contributors import SuiteContributorByModule
+from corehq.apps.app_manager.helpers.make_build.suite_xml.xml_models import Menu, Command, LocalizedMenu
 from corehq.apps.app_manager.util import (is_usercase_in_use, xpath_references_case,
     xpath_references_user_case)
 from corehq.apps.app_manager.xpath import (interpolate_xpath, CaseIDXPath, session_var,
