+++ corehq/apps/app_manager/helpers/make_build/suite_xml/sections/resources.py
@@ -4,8 +4,8 @@ from __future__ import unicode_literals
 from distutils.version import LooseVersion
 
 from corehq.apps.app_manager.helpers.make_build import id_strings
-from corehq.apps.app_manager.suite_xml.contributors import SectionContributor
-from corehq.apps.app_manager.suite_xml.xml_models import LocaleResource, XFormResource, \
+from corehq.apps.app_manager.helpers.make_build.suite_xml.contributors import SectionContributor
+from corehq.apps.app_manager.helpers.make_build.suite_xml.xml_models import LocaleResource, XFormResource, \
     PracticeUserRestoreResource, ReleaseInfoXFormResource
 from corehq.apps.app_manager.templatetags.xforms_extras import trans
 from corehq.apps.app_manager.util import languages_mapping
