+++ corehq/apps/app_manager/helpers/make_build/suite_xml/utils.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 from lxml import etree
 from django.utils.translation import ugettext as _
 from corehq.apps.app_manager.exceptions import SuiteValidationError
-from corehq.apps.app_manager.suite_xml.xml_models import Suite
+from corehq.apps.app_manager.helpers.make_build.suite_xml.xml_models import Suite
 import six
